+++ example/lib/bloc/login_event.dart
@@ -1,0 +1,7 @@
+sealed class LoginEvent {}
+
+class LoginSubmitted extends LoginEvent {
+  LoginSubmitted(this.username, this.password);
+  final String username;
+  final String password;
+}