+++ lib/flutter_bloc_effects.dart
@@ -1,3 +1,3 @@
-export 'bloc_effects.dart';
+export 'bloc_effect_emitter.dart';
 export 'bloc_effect_listener.dart';
 export 'multi_bloc_effect_listener.dart';